create task tile
